Rearrange the folder structure to be a library
This repository is advertised as a "library for Arduino" but the previous folder structure makes it a sketch, not a library. Rearranging the folder structure allows installation as a library after which this library can be shared between multiple projects by simply #include-ing the header file. The example sketch will appear under File > Examples > libSNMP in the Arduino IDE.